chore: remove all possible AsString panics (#183)
Any call to `AsString` is unsafe. All places where the error is ignored
is used in errors, so a placeholder or empty string is ok.

A panic was observed by a customer. I am not sure where it occured. To
be safe, I am using the function we have to protect these `AsString`
calls. I added a custom linter to prevent these calls, this PR just
removes all calls that were not fixed already.